requirements.txt
pillow
numpy
pygame
pyqt6-sip
